[REFACTOR] change talentPool into talentpool
