--- src/main/java/rencanakan/id/talentPool/TalentPoolApplication.java
@@ -1,14 +1,0 @@
-package rencanakan.id.talentPool;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-
-@SpringBootApplication()
-public class TalentPoolApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TalentPoolApplication.class, args);
-	}
-
-}
