--- src/main/java/rencanakan/id/talentPool/controller/ExperienceController.java
@@ -1,24 +1,0 @@
-package rencanakan.id.talentPool.controller;
-
-import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import rencanakan.id.talentPool.dto.EditExperienceRequestDTO;
-import rencanakan.id.talentPool.dto.ExperienceResponseDTO;
-import rencanakan.id.talentPool.dto.WebResponse;
-import rencanakan.id.talentPool.service.ExperienceService;
-
-@RestController
-@RequestMapping("/api/experiences")
-public class ExperienceController {
-    @Autowired
-    ExperienceService experienceService;
-
-    @PutMapping("/{id}")
-    public WebResponse<ExperienceResponseDTO> editExperienceById(@PathVariable Long id, @RequestHeader("Authorization") String token,   @Valid @RequestBody EditExperienceRequestDTO dto) {
-        ExperienceResponseDTO resp= experienceService.editById(id, dto);
-        return WebResponse.<ExperienceResponseDTO>builder().data(resp).build();
-    }
-
-
-}
