+++ src/main/java/rencanakan/id/talentpool/TalentPoolApplication.java
@@ -1,0 +1,13 @@
+package rencanakan.id.talentpool;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication()
+public class TalentPoolApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TalentPoolApplication.class, args);
+	}
+
+}
