+++ src/main/java/rencanakan/id/talentpool/dto/ExperienceResponseDTO.java
@@ -1,0 +1,23 @@
+package rencanakan.id.talentpool.dto;
+
+import rencanakan.id.talentpool.enums.EmploymentType;
+import rencanakan.id.talentpool.enums.LocationType;
+
+import java.time.LocalDate;
+import lombok.*;
+
+@NoArgsConstructor
+@Getter
+@Setter
+@AllArgsConstructor
+public class ExperienceResponseDTO {
+    private Long id;
+    private String title;
+    private String company;
+    private EmploymentType employmentType;
+    private LocalDate startDate;
+    private LocalDate endDate;
+    private String location;
+    private LocationType locationType;
+    private Long talentId;
+}
