+++ src/main/java/rencanakan/id/talentpool/dto/WebResponse.java
@@ -1,0 +1,15 @@
+package rencanakan.id.talentpool.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class WebResponse <T> {
+
+    private T data;
+
+    private String errors;
+}
