+++ src/main/java/rencanakan/id/talentpool/service/ExperienceService.java
@@ -1,0 +1,8 @@
+package rencanakan.id.talentpool.service;
+
+import rencanakan.id.talentpool.dto.EditExperienceRequestDTO;
+import rencanakan.id.talentpool.dto.ExperienceResponseDTO;
+
+public interface ExperienceService{
+    ExperienceResponseDTO editById(Long id, EditExperienceRequestDTO dto);
+}
