+++ src/test/java/rencanakan/id/talentpool/TalentPoolApplicationTests.java
@@ -1,0 +1,15 @@
+package rencanakan.id.talentpool;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class TalentPoolApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
